Class Filme

+++ CatalagoFilmes/CatalagoFilmes/src/tech/clairtonlima/t1113/model/Filme.java
@@ -1,12 +1,11 @@
 package tech.clairtonlima.t1113.model;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Filme {
     private String nome;
     private LocalDate dataLancamento;
-    private BigDecimal orcamento;
+    private double orcamento;
     private  String descricao;
     private List<Ator> atores;
     private List<Diretor> diretores;
